refactor: extract Form_Records_Meta_Box from Form_Submissions
Move the single-record edit screen (submission-data meta box, Update box with
delivery status, admin-menu placement, field save handler) into a dedicated
Form_Records_Meta_Box class registering its own add_meta_boxes/admin_menu/
save_post hooks. Add a characterization test for saving edited field values
via the save_post hook.
