+++ phpstan-baseline.neon
@@ -1795,10 +1795,10 @@ parameters:
 			path: inc/plugins/class-form-submissions.php
 
 		-
-			rawMessage: 'Method ThemeIsle\GutenbergBlocks\Plugins\Form_Submissions::render_field() has parameter $field with no value type specified in iterable type array.'
+			rawMessage: 'Method ThemeIsle\GutenbergBlocks\Plugins\Form_Records_Meta_Box::render_field() has parameter $field with no value type specified in iterable type array.'
 			identifier: missingType.iterableValue
 			count: 1
-			path: inc/plugins/class-form-submissions.php
+			path: inc/plugins/class-form-records-meta-box.php
 
 		-
 			rawMessage: 'Method ThemeIsle\GutenbergBlocks\Plugins\Form_Submissions::update_submission_dump_data() has no return type specified.'
